Restored ability to pass in repos
The only thing I changed was the name, instead of the former name repos, I chose
extra-repos to reduce confusion in the code. An easy tweak for cljdoc at
integration time.

Also moved default repos to our config.edn, they are static, and this seemed a
better place for them.

Fixes #12