Package metagetta subproject as a jar
We currently consider metagetta to be an internal subproject of cljdoc-analyzer.

When testing how cljdoc-analyzer itself is analyzed by cljdoc, a failure
occurred because the analyzer tries to parse all source and metagetta source was
not in a location that was loadable (namespace did not match path).

To truly treat metagetta as a black box resource, it is now included in the
cljdoc-analyzer jar as metagetta.jar.

After this change cljdoc could analyze cljdoc-analyzer just fine. This allowed
me to review and fixup current api docs and some fn naming.