Compute all js dependencies used by package in early stage.
This allow the good execution of "analyze-file" on any file of the package.
In the previous behavior, if one on the embedded cljs dependencies has a js
dependency, this dependency is not detected, so the call to "analyze-file"
fails.

To avoid that, every js dependencies have to compute in early stage and
pass downstream.