Add failing test for unknown tagged literal
for #32

NOTE: I added a separate test source dir because adding it
to the existing one made most of the other tests fail as well,
which is an unfortunate coupling. The new test-sources-special
is intended for testing 1 ns at a time so adding a new ns and test
won't break existing ones.