Process .cljs first when analyzing for cljs
CI was failing integration tests on the orchestra library.
The same tests passed for me locally on my dev box.
Returned analysis had a namespace with .cljc vars but it was missing
.cljs file vars.

The issue was ClojureScript analysis file processing order was
undefined. It is important to process .cljs files before .cljc
files. My dev box happened to deliver the files in the happy order.
CI delivered them in an unhappy order.

Now processing files in a deterministic order for ClojureScript
analyis where .cljs files will be processed before .cljc files.