Add tests for new no-doc load behavior
We now verify:
1. When a namespace is not loadable due to some error, but marked with
   no-doc metadata, because the this namespace is no longer loaded, the
   analysis still passes.
2. When a namespace is marked dynamically at load time with no-doc
   metadata, the namespace is skipped.

Test setup for item 1 was a bit awkward.
We only want to include unloadable namespaces when testing with no-doc
exclusion options. Currently a special source root is automatcially
setup under the target dir for this test.

When verifying item 2, I discovered that we were ignoring any load
time alterations to namespace metadata for ClojureScript.
Altering at load-time is somewhat commonly used in Clojure
(potemkin import-vars, etc), but probably not ClojureScript.
Just the same I corrected the oversight with this change.

Closes #45