Be less naive when parsing poms
Use tools deps which uses Maven libs to parse and evaluate pom.

We are using tools deps internals.
If that becomes a maintenance headache, we will adapt, but for now it is
convenient.

Other than parsing poms via maven, the algorithm remains unchanged, we
still, for example:
- only look at non-transitive deps
- do not consider dep exclusions for provided deps

Added an integration test for org.clojure/tools.deps to cover one former
problematic use case.

We'll adjust if needed, but might not need to.

Closes #66
Closes cljdoc/cljdoc#742