Support analysis of Clojure 1.10+
Added support necessary to successfully analyze Clojure v1.10+
- jar cleanup now removes clojure generated .class files for both proxies and types
- exclude clojure.parallel from analysis

Contributes to #53

The :namespaces option is was applied after analysis. This meant that
namespaces that were to be excluded because they would cause analysis to fail
still caused analysis to fail. Moved exclusion logic to after namespaces
are found but before analysis.

Closes #92