Support analysis of Clojure 1.7+
Metagetta assumed current and compatible versions of Clojure and
ClojureScript. When analyzing older versions of Clojure we can pick up a
version that is not compatible with our default ClojureScript.

Since we don't need ClojureScript to analyze Clojure, only bring in
its support when it is needed.

Also: clojure.string/starts-with? was introduced in Clojure 1.8.
Use JVM .startsWith instead.

Contributes to #53