Exclude ns x in .cljc source if defined in .clj
For Clojure analysis, when finding namespaces to process, exclude ns `x`
defined in `.cljc` source if it is already defined in a `.clj` source.

This replicates Clojure v1.6+ ns loading and fixes duplicate ns issue
for libs like instaparse.

Patched inlined tools.namespace 1.4.0 to bring in feature (and fix for):
https://clojure.atlassian.net/jira/software/c/projects/TNS/issues/TNS-61

Closes #101